# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Ledge API",
        "version": "1.0.0"
    },
    "components": {
        "schemas": {
...
            "BadRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "BAD_REQUEST"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
...
            "PayloadTooLarge": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "PAYLOAD_TOO_LARGE"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
            "UnprocessableContent": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "UNPROCESSABLE_CONTENT"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
            "TooManyRequests": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "TOO_MANY_REQUESTS"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
            "InternalServerError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "INTERNAL_SERVER_ERROR"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
...
            "Unauthorized": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "UNAUTHORIZED"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
...
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "value": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "income",
                            "expense"
                        ]
                    },
                    "category": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "need",
                                    "want",
                                    "investment"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "date": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "userId",
                    "name",
                    "value",
                    "type",
                    "category",
                    "date",
                    "createdAt",
                    "updatedAt"
                ],
                "additionalProperties": false
            },
            "Forbidden": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "FORBIDDEN"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            },
            "TransactionNotFound": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "const": "TRANSACTION_NOT_FOUND"
                    }
                },
                "required": [
                    "code"
                ],
                "additionalProperties": false
            }
        }
    },
    "paths": {
...
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionNotFound"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TooManyRequests"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "number"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "income",
                                            "expense"
                                        ]
                                    },
                                    "category": {
                                        "anyOf": [
                                            {
                                                "type": "string",
                                                "enum": [
                                                    "need",
                                                    "want",
                                                    "investment"
                                                ]
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                                    }
                                },
                                "required": [
                                    "name",
                                    "value",
                                    "type",
                                    "category",
                                    "date"
                                ]
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionNotFound"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PayloadTooLarge"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContent"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TooManyRequests"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted successfully"
                    },
                    "400": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionNotFound"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PayloadTooLarge"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TooManyRequests"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}